Adapted framework to be usable with both python 2 and 3.
The addition of python 2 functionality allowed for the addition of
a ROS turtlebot line follower, and it actually works!  It follows
anything that's yellow.  Gazebo unfortunately does not allow for
dynamic lines to be drawn in the world, so I buried some yellow
dividers in the ground.  I need to figure out something better,
because the turtlebot gets stuck on them since they're 3d.